Fix failing tests by renaming "1.7" to "java8"
We do not resolve the platform name "1.7" to "java8", even though it seems we're supposed to be doing this. So, Python 3 was always failing. With the most recent commit to tighten the failure expected error message, this meant Py2 was also failing.

Maybe we should support this feature? I'll open a separate issue to track that. For now, we focus on fixing the tests.