Remove option integration test from Python 3 blacklist (#7142)
I'm not sure why it was added to begin with. Even back in #6959, which first introduced `./pants3`, the tests were passing.

Maybe it was a flaky test? But I can't find a ticket for that. So maybe it was a typo?

--

This also redistributes CI shards by one in favor of Python 3, as this and the surrounding PRs remove multiple targets from the blacklist.