Make Py interpreter->PEX digest lookup more robust
Uses {}.get() and a None check with a raise to avoid a potentially bad KeyError. It's not very likely we would get a KeyError to begin with, as this would mean they're running Pants with Python 2.6 or 3.3-3.5, but it does help future developers to better understand the intent of this code.

Note I wasn't certain which Exception to raise. I considered UnsatisfiableInterpreterConstraintsError, but felt it was meant to be used in a different context. So I simply used OSError. Happy to change this to something more relevant to our domain.