Use ValueError instead of OSError
OSError is supposed to be reserved for system calls failing to function. It is not the same thing as your external system not being setup properly (such as using the wrong Python version).

Instead, ValueError is more appropriate. Even better would be our own custom error, but I could not find a preexisting error to use and this branch is unlikely to happen (only if the user is using Py 2.6, 3.3-3.5, or 3.8) so the ValueError suffices.

Thanks Daniel for the suggestion!