Don't re-panic in mock execution server
Right now, if a test fails _and_ didn't make the right requests, we get
a really unhelpful message (about a thread panicking while unwinding
panick handlers) which hides the actual assertion message.

Instead, don't re-panic if the thread is already unwinding a panic, just
print out the error it would've printed.

If the thread _isn't_ unwinding a panic, we still panic, because we
still need to fail the test.