Properly decode Go Compile env variables
Failing to decode them was causing the byte string prefix `b''` to pop up in filenames, which led to test failures. For example, `GOOS="windows" ./pants3 binary contrib/go/examples/src/go/hello` would fail that it could not find the file, because the filename was nonsense with b'' in it.