Fix JVM tests issue comparing None to str
We call sorted() on a tuple of 6 properties taken from JUnitTests. Some of these properties may be None if the value is missing, but sometimes they have an actual value. Because sorted() calls __lt__(), comparisons between these missing and valid properties result in a TypeError.

So, implement a custom comparator.

Maybe this should be moved to an __lt__() function for JUnitTests()..?