Start testing pre commit hooks (#7118)
### Problem

Begins to address #7069. Our pre-commit hooks are nice, and also completely untested, leading to failures which caused the reversion of #7013 in #7028 (and that change is still necessary to run python checkstyle on my osx machine).

I wasn't able to get to all of it, but I was able to test some of the trickier logic introduced in #7066 (e.g. operating on newly added files specifically, and checking the current copyright year) in `test_git.py`.

### Solution

- Pass in the directories to check as `$@` to `check_packages.sh` and `check_header_helper.py` so that we can DRY declare them in `pre-commit.sh`.
  - Remove `check_header.sh` (step 0 to making everything in python! :tada:).
- Add unit tests for both of those scripts.
- Add a unit test for finding newly added files, which was moved into a new script `get_added_files.sh` instead of inlined in another script.
- Change `IGNORE_ADDED_FILES` env var to reduce copyright year checks to `PANTS_IGNORE_ADDED_FILES` on a suggestion from @Eric-Arellano to ease the path into more pythonification :rocket: 

### Result

We now have a platform to iterate on logic or performance changes to pre-commit hooks without introducing random breakage.