Fix typo in ci.sh: 3.6, not 3.7
Note this does not actually change any tests. Here, `py_version_number` solely changes what is printed to the screen with `banner`.

At the moment, we consistently use Py3.6 in every Py3 shard, so this more constrained message of saying "Using Python 3.6" instead of using "Python 3" is correct. The motivation is pre-work for the Py37 in CI pr.

Happy to take out need be.

Thanks for the catch, John!