Change part 2 of solution to use $PY env var instead of ./pants3 flags
The original solution of using `./pants -6` and `./pants3 -7` was overcomplicated and overconstrained.

Instead, per John's suggestion, this uses the $PY env var that we were using in `virtualenv` to achieve the same end goal in a much simpler way. If the user wants to constrain to something more specific than `python3`, they can do so with `PY=python3.7 ./pants3`.