fetch native libraries from python requirements, e.g. tensorflow to define custom operators (#7046)
### Problem

We can depend on 3rdparty Conan packages wrapping C and C++ libraries in our native backend compile and link tasks. This works for packages that are distributed this way, but until a build tool swoops in and saves the day, it's difficult to develop an ecosystem of 3rdparty C/C++ packages from scratch. In particular, TensorFlow forgoes this approach in favor of packaging their native libraries directly in their python wheel. This works fine if using the soon-to-be-removed `python_dist()` support for distutils `Extension`s (see #7016), but without that, a separate Conan package for TensorFlow must be created and kept up to date with the python requirement.

A variety of python packages distribute native code this way, but TensorFlow in particular allows defining custom operators, which allow hooking deeply into its internals for greater performance on specific tasks: see https://www.tensorflow.org/guide/extend/op. The tutorial there contains a complete example for defining custom operators in Bazel directly from the TensorFlow source tree, and gives clear enough directions on how to define custom operators manually, but those instructions do not generalize to other python packages which may distribute native code embedded in the dist. The utility of extracting code directly from a python requirement means that C and C++ targets which depend on this `native_python_wheel()` target which wraps a `python_requirement_library()` containing TensorFlow can expect to always be in sync with python code which depends on the requirement library.

If TensorFlow were to start packaging its code separately from the python wheel, then we might switch back to a Conan package or a Conan recipe, but as of now, this is a nice first step to synchronizing dependencies across languages in a more structured way than just making sure the version strings are in sync (which doesn't always break things if you forget). I believe this approach is highly generalizable and extremely useful for allowing extremely low-mental-overhead interop between C/C++ and other languages by bridging the gap between user code and native libraries that are only available through other package managers (npm and CRAN are some other immediate examples).

### Solution

- Introduce `unpacked_whls()` which specifies how to pull native code our of a `python_requirement_library()` target, and can be used as a source target for `remote_sources()` to produce a `packaged_native_library()` target as per #7060.
- Add an `UnpackWheels` task which performs the extraction specified by an `unpacked_whls()` target, enabled by the work to split out a `UnpackRemoteSourcesBase` superclass in #7061.
- Implement https://www.tensorflow.org/guide/extend/op entirely in pants using `unpacked_whls()` and `remote_sources()`.

### Result

Users have a clear example of how to write custom operators for TensorFlow in pants with a managed native toolchain, which doesn't require them to build from the TensorFlow source tree, unlike the Bazel example in https://www.tensorflow.org/guide/extend/op. This extraction functionality allows deeply integrating into python packages which expose native code in a hygienic way which doesn't require any new concepts or custom wrappers. We extend the `remote_sources()` concept to cover python wheels, immediately opening up a number of new ideas for how to integrate the smorgasbord of native code available on the internet.