Add PY= arg to ci.sh when using Py2
Without this line, the command `./build-support/bin/ci.sh -e2` will fail because $PY is undefined. We don't use this command in CI at the moment, but it should work.

The reason we need to add this line is that we don't bootstrap the PEX with that command nor do we call `./pants`, so we do not set the $PY variable anywhere, which leads to it being undefined and the code in `pants_venv` failing.