Change OSX rust tests to use Py2
Although the tests does not care about which Python version it uses, it for some reason is invoking `pants_venv`, which is complaining that it cannot find python3.6.

We did not discover this until now because we earlier simply used the bin `python3`, so whatever was on the system (e.g. 3.4 or 3.5) would resolve without issue. Now, we require 3.6 to operate correctly.

As we don't care which version we use, we can just change the test to use Py2.