Move cli flag parsing to the top of the file
This allows us to use the results of the command line flags right away throughout the rest of the script, whereas earlier we could only access the results in functions that were evaluated after the global code executed.

Without this change, we won't be able to read the flag `-3` and use it to set the `PY` env var.

There's precedent for this in how ci.sh sets up its cli parsing code at the top.

Note we must now as a result introduce variables called run_x to keep track of what the user wants to run, and then at the bottom of the script read those values to execute the commands.