Refactor packages.py in preparation for adding Python 3 support (#7190)
### Problem
While working on adding Python 3 support to `release.sh`, the command line parsing in `packages.py` became confusing in trying to support an optional flag `--py3`. The current implementation of `sys.argv[1:]` works at the moment, but is difficult to work with for more complex command line configurations.

### Solution
Use argparse, which is Python's official library for CLI flags.

Focus on preserving the original semantics, while also making the code more readable and option parsing more robust.

### Result
Script functions the same as before and is ready for adding Python 3 support.