Add flag reporting-zipkin-sample-rate (#7211)
### Problem
In the current implementation, every time the pants command is run with zipkin tracing turned on all the zipkin traces will be collected. It is not very convenient when the number of runs is very big. 

### Solution
Possibility to set the sample rate will allow us to have the number of traces that fits the constraints of Zipkin server. 

### Result
A flag `reporting-zipkin-sample-rate` was added that sets the sample rate at which to sample Zipkin traces. If flags `reporting-zipkin-trace-id` and `reporting-zipkin-parent-id` are set the sample rate will always be 100.0 (no matter what is set in `reporting-zipkin-sample-rate` flag).