Add checks if values of flags zipkin-trace-id and zipkin-parent-id are valid (#7242)
### Problem
When pants are called with flags zipkin-trace-id and zipkin-parent-id an assertion error is raised if the values of the flag are of the wrong format. The error is not informative. 

### Solution
Checks of values of flags zipkin-trace-id and zipkin-parent-id are added with a better error explanation. Users of the pants are asked to use 16-character  or 32-character hex string.
Also, tests are added for these checks.