Scheduler returns metrics as a dictionary instead of a tuple of tuples (#7255)
### Problem
Scheduler returns metrics as a tuple of tuples (key, value). And later this tuple is transformed into a dictionary.
It is considered to use metrics to return to python part zipkin span info and dict type will be more convenient.

### Solution
Scheduler_metrics returns a store_dict  instead of store_tuple.