Deduplicate OSX Pyenv and OpenSSL env var
We now have to set up this env var 4 distinct times. With https://github.com/pantsbuild/pants/pull/7235, we will have to again use the env var in order to install Python 2 with UCS4.

So, this code needs to be de-duplicated. We keep the name generic to any Python interpreter because it's really about Pyenv, not Python 3.