[rsc-compile] bump version; rm metacp jobs; update tests
Rsc has been updated so that it can consume regular classpaths in all cases, eliminating the need to run metacp before and after rsc. This removes all the extra jobs. It is based on @cosmicexplorer's https://github.com/pantsbuild/pants/pull/7227.

The compile portions are ready for a first pass of review, but I still need to do more work on the tests.

Travis will fail on it right now, because the newer Rsc version hasn't been pushed to maven central yet.