Revert "Constrain ci.sh to the exact Python interpreter version"
This reverts commit 887a8ef035a81b79826f7c3b93231586b8a07c10.

This change is necessary to fix the original motivation for this PR, but it does not really belong in this PR anymore. Instead, it should be in the Py2 ABI PR (7235). This PR should be kept more generic, and there is no logical connection to the changes being made with ci.sh, beyond that original motivating problem.