Allow tasks to opt-in to target filtering (#7283)
Followup to https://github.com/pantsbuild/pants/pull/7275 following [discussion](https://github.com/pantsbuild/pants/pull/7275#discussion_r259557791) that the target filter was being applied to tasks that do not support it (e.g. tasks that don't access targets via `get_targets()`)

This adds a class property to `Task` that allows subclasses to effectively opt-in to the new behavior - and sets that flag to `True` for `fmt` and `lint` tasks.