Use Pex36, not Pex37
Originally we did not release Pex36 so the only option was Pex37. This was an issue because Pex37 means we need a Py37 interpreter available, which we do not have in the Py3 wheel building shards and do not want to setup as it would require bootstrapping Pants again.

Note this commit is going to be blocked on the upgrade to pex 1.6.2 to work properly. But once that's merged, this should work.