Ignore Pex deprecation warnings
I've been waiting to land this PR for the upgrade to Pex 1.6.* as to avoid a regex DeprecationWarning anytime you run with Py37. That upgrade ran into an issue that needs to be broken out into a new Pex release.

In the meantime, we can simply ignore the DeprecationWarning and remove this new line as a part of the Pex upgrade PR.

Will merge this PR after a green CI run!