Cherry-pick CI changes from #7235
I realized the original ci.sh would not work as intended! We would always hardcode PANTS_PYTHON_SETUP_INTERPRETER_CONSTRAINTS with `-2` to be >2.7,<4. This means our blacklist shard would not be able to properly constrain to >=3.6, and would actually be running with 2.7. The tests were all green, so we didn't notice this.

Instead, we copy the code from #7235 (py2 abi specified), as I've iterated on ci.sh a couple distinct times there until finally discovering this as the best version. These changes make sense to have here, because we do need to change how we configure the interpreter constraints, so might as well do it right the first time here.