Move OpenSSL and Py37 install to new Dockerfile
Setting up OpenSSL takes a long time. Same with Py37. We only need to do either of these once a day though, at night for that first Linux Py37 bootstrap shard. So, even though it leads to duplication, split this out into a new Dockerfile.

Eventually, this file should be deleted as the Centos6 image should come with Py36 and Py37 preinstalled.