Add libffi to Linux shards
All Py37 linux test shards are failing with the message `libffi.so.5: cannot open shared object file: No such file or directory`. This message appears when libbfi is not installed in the environment, and has appeared with many other libraries if you simply search that error.

I originally thought this was an issue with the Docker image, but realized it isn't because we already install libffi there. Instead, I found we don't install libffi for Linux shards. I'm surprised Py27 and Py36 have been working given that!