Split out rsc distribution selection into JvmCompile (#7290)
### Problem

See https://github.com/pantsbuild/pants/pull/7272#discussion_r259156484. We are selecting the jvm distribution in a pretty manual way in `rsc_compile.py` that might benefit from being extracted into the `JvmCompile` base class.

### Solution

- Move `_get_jvm_distribution()` from `RscCompile` into `JvmCompile`.
- Use that logic to simplify `ZincCompile#_get_zinc_arguments()`.

### Result

Distribution selection, and in particular preparing the distribution to be usable in a hermetic context, is extracted into a common method in the jvm compile base class.