Default to pants devs using Python 3 by renaming `./pants3`->`./pants` and `./pants`->`./pants2` (#7257)
### Problem
We should internally default to using Python 3, as we will in two major releases require Python 3 to run.

Further, we want to make this change before officially releasing a Python 3 wheel so that we can internally dogfood the migration before others consume it.

### Solution
Rename `./pants3` -> `./pants` and `./pants` -> `./pants2`. This means `./pants` will default to using Python 3, and `./pants2` will default to using Python 2.

Also update `ci.sh` to properly set `PANTS_PYTHON_SETUP_INTERPRETER_CONSTRAINTS`. We now allow the caller of the script to pre-set the value (along with `$PY`), and default to always setting the constraint to the exact Python interpreter used for `$PY`. This allows us to ensure we always use the correct interpreter, while allowing certain tests to override the values, such as the Py2 blacklist shard to use a Py2 Pex with Py3 interpreter constraints.

Note for both CI and for the `./pants` script, we set interpreter constraints to the exact same version as used by `$PY`, even including the same patch version. This ensures the interpreter does not unintentionally change between what we use for Pants vs. for spawned subprocesses. The user can always override both `$PY` and the interpreter constraints, so the constraints can be loosened when desired.

### Result
* Running `./pants` now defaults to unconstrained Python 3, i.e. Python 3.6 or 3.7.
* Running `PY=python3.7 ./pants` runs constrained Python 3.7. Same with `PY=python3.6` running Python 3.6.
* Running `./pants2` defaults to running Python 2.7 with interpreter constraints allowing for either Python 2 or Python 3 for spawned subprocesses such as unit tests.

#### Not impacted by change
Note this does not impact end consumers of Pants, only Pants developers working directly on this repository.