filespec matching utility functions use native glob matching. (#7299)
### Problem

As described in #6795, we have python code with functionality that duplicates rust `PathGlob` matching, which is undesirable. So undesirable, in fact, that we recently noticed a bug where `--changed` and `--owner-of` (using that code) would not correct identify a file as being owned (globs like `./*.py` would never match).

### Solution

`PathGlob` expansion is currently implemented by lazily traversing a filesystem, and it is unclear how to decouple the lazy traversal from the glob matching without double-matching paths (once to implement the laziness, and then again to accept a complete path). Additionally, the codepath we are replacing is not particularly performance sensitive.

So: this patch takes the approach of implementing a small `impl VFS for MemFS` against which to expand `PathGlobs`. This means we pay the cost of the laziness, but it ensures an identical implementation of glob matching (assuming that the small `MemFS` is correct).

### Result

Fixes #6795, and ensures that matching `PathGlobs` against a set of in memory file paths has identical behaviour to expanding against a filesystem.