move the declarative task stuff out of the python backend testing (#7279)
### Problem

We want to do #7128, which involves adding some more C/C++ unit testing. We would love to use some of the declarative task specification stuff we've added to the `python_dist()` testing to do that.

### Solution

- Move the generic logic out of `BuildLocalPythonDistributionsTestBase` into `DeclarativeTaskTestMixin` in `task_test_base.py`.

### Result

A new interface for task unit testing (which can be extended to more than just v1 tasks) is available to other testing.