add a lint to warn on implicit string concatenation (#7286)
### Problem

It's occasionally possible to make mistakes when formatting strings using the "implicit" syntax of just placing two string literals next to each other. Although this pattern is used extremely commonly in the codebase right now, it would be nice to be able to point this out.

### Solution

- Add a linter for string concatenations of the form e.g. `'a' 'b'`.
- Add a suppressions file which turns it off for all python files in this repo.

### Result

It's possible this could lead to fewer errors. However, there is currently a lot of noise induced by turning this on and it would be nice to avoid that by eventually fixing code to remove it (e.g. by using triple-quoted `"""` strings instead), but for now, it has been suppressed. Other repos can freely use this lint, however.