Filter out irrelevant env vars in OptionsBootstrapper. (#7312)
### Problem

The `OptionsBootstrapper` holds the env/args/config that can eventually be fully parsed into an `Options` object that can be consumed by the rest of pants.

Options parsing only consumes `env` vars that start with `PANTS_`, but the `OptionsBootstrapper` currently captures the entire `env`. This can make the `OptionsBootstrapper` unnecessarily large (and its `__str__` unwieldy).

### Solution

Filter env vars recorded in the `OptionsBootstrapper` to only those starting with `PANTS_`.