@union / UnionRule for letting the engine figure out paths to products not known in advance (#7116)
### Problem

*Resolves #4535.*

It's currently difficult for rules to tell the engine "give me an instance of `y` from this `x` according to the registered rules in the rule graph" if `x` is not a specific type known in advance (see #4535 for use cases for this). The alternative is that upstream rules have to use conditional logic outside of the engine, which is very difficult to write in a way that combines with arbitrary rulesets from downstream plugins, and code written to get around this can be error-prone. This is sad because the rule graph is generated by combining rules in pants core as well as in plugins, so all the necessary information is there, we just can't make use of it.

This PR introduces the `@union` decorator and `UnionRule` type to allow for static resolution of "union" types in `yield Get(Product, UnionType, subject)` statements as per [this comment in #4535 and below](https://github.com/pantsbuild/pants/issues/4535#issuecomment-455676940).

### Solution

#### Python
- Add a `subject_declared_type` field to `pants.engine.selectors.Get` instead of inferring it from the subject type (the 2-argument form of `Get` is still supported).
- Introduce `Get.create_statically_from_rule_graph()` classmethod to make it clear when the `Get` subject is populated or not.
- Introduce `@union` and `UnionRule` to describe union types which can be requested with a `yield Get(Product, UnionType, subject)` in rule bodies (as per [this comment on #4535](https://github.com/pantsbuild/pants/issues/4535#issuecomment-455676940)).
  - **Note that `@union` classes are not registered in `def rules(): ...` -- this distinction seems to make sense as union classes are never instantiated.**
- Create a really simple `union_rules` dict field in `RuleIndex` which registers `@union_rule()` decorators as a map from `union base type -> [union members]`.
- Propagate the `union_rules` dict to the scheduler, and when adding `Get` edges (in `_register_task()` in `scheduler.py`), check if the `subject_declared_type` is a union base, and if so add edges to all union members.
- Create a `HydrateableField` union base which is used to resolve (for now) either `SourcesField` or `BundlesField` in a `yield Get()`.

### Result

Users can now dynamically add union members in plugins and backends to be processed by upstream rules using `yield Get(...)` which don't know anything about them, and with a static universe of known union members which the engine uses uses to type-check the subject of a  `yield Get(...)` at rule execution time.