Print trace id to console and possible URL with the zipkin trace (#7313)
### Problem
It is not obvious for the user how to find the zipkin trace after a pants run.

### Solution
One of the solutions is to print zipkin trace in the console. 
The better one could be to print a link to the Zipkin API with the trace. But endpoints to send spans to Zipkin API and to view the resulting trace may be different. The set up is not unique. 
One of the solutions is to have two flags:
--reporter-zipkin-send-endpoint  and
--reporter-zipkin-view-endpoint
Current implementation suggests a link that can point to the current zipkin trace.

I would be grateful for your comments in regard to it